+++ src/opts/mod.rs
@@ -4,6 +4,7 @@ mod containers;
 mod exec;
 mod images;
 mod manifests;
+mod networks;
 mod pods;
 mod volumes;
 
@@ -11,6 +12,7 @@ pub use containers::*;
 pub use exec::*;
 pub use images::*;
 pub use manifests::*;
+pub use networks::*;
 pub use pods::*;
 pub use volumes::*;