+++ CHANGELOG.md
@@ -20,3 +20,4 @@
 * Add `Manifest::remove_image`
 * Add `Manifest::push`
 * Add `Network::delete` and `Network::force_delete`
+* Add `Networks::create`